+++ libpurple/protocols/msn/transaction.h
@@ -75,7 +75,7 @@ char *msn_transaction_to_string(MsnTransaction *trans);
 void msn_transaction_queue_cmd(MsnTransaction *trans, MsnCommand *cmd);
 void msn_transaction_unqueue_cmd(MsnTransaction *trans, MsnCmdProc *cmdproc);
 void msn_transaction_set_payload(MsnTransaction *trans,
-								 const char *payload, int payload_len);
+								 const char *payload, gsize payload_len);
 void msn_transaction_set_data(MsnTransaction *trans, void *data);
 void msn_transaction_set_data_free(MsnTransaction *trans, GDestroyNotify fn);
 void msn_transaction_set_saveable(MsnTransaction  *trans, gboolean saveable);
